==> notebooks/0.01-md-data_cleaning.ipynb <==
import pandas as pd

# Load the dataset
data = pd.read_csv('../data/external/matches.csv')
data

# Certain dates and seasons are inconsistent - matches from year 2020 are labeled as season 2024!

# Create copy of the raw data and drop unnecessary columns
data = data.drop(labels=['Unnamed: 0', 'time', 'comp', 'day', 'venue', 'attendance', 'captain', 'referee', 'match report', 'notes'], axis=1)

# Explore the data types
data.info()

#Check for missing values
data.isna().sum()

#Fill NaN values with column's mean
data['dist'] = data['dist'].fillna(value=data['dist'].mean())

#Check for missing values again
data.isna().sum()

# Summarize the dataset with descriptive statistics
data.describe()

# Check season inconsistency
data['season'].value_counts()

# Drop duplicate rows while retaining the last occurrence (most recent data)
subset = data.drop(labels='season', axis=1)
data.drop_duplicates(subset=subset, keep='last', inplace=True)

# Re-check season consistency after cleaning
data['season'].value_counts()

# Correct data types for categorical variables and date
data[['result', 'opponent', 'team']] = data[['result', 'opponent', 'team']].astype('category')
data['date'] = pd.to_datetime(data['date'])
data.info()

#Export DataFrame to a csv file
data.to_csv('../data/interim/data_clean.csv', index=False)
